questao 1 concluida

+++ listas/lista_01/lista_01_questao_1.c
@@ -1,11 +1,14 @@
 #include <stdio.h>
 
 // URI Online Judge - Problema 1176 - Fibonacci Array - Nível 4
+// Aluno: Rodrigo Matos Aguiar - Rudigus
+
+// Não use unsigned long nessa questão. Dependendo da máquina, tanto o unsigned long como o unsigned long long podem armazenar 8 bytes ou 64 bits. Todavia, isso não vale pra todas as máquinas. No URI, o unsigned long armazena 4 bytes, então não irá suportar termos de Fibonacci elevados o suficiente.
 
 // Função que recebe como parâmetro um número N e retorna o N-ésimo termo da sequência de Fibonacci.
-unsigned long fibonacci(int numFornecido)
+unsigned long long fibonacci(int numFornecido)
 {
-	unsigned long n1 = 0, n2 = 1, aux = 0;
+	unsigned long long n1 = 0, n2 = 1, aux = 0;
 	for(int i = 0; i < numFornecido; i++)
 	{
 		aux = n2;
@@ -24,6 +27,6 @@ int main()
 	{
 		int numFornecido;
 		scanf("%d", &numFornecido);
-		printf("Fib(%d) = %lu\n", numFornecido, fibonacci(numFornecido));
+		printf("Fib(%d) = %llu\n", numFornecido, fibonacci(numFornecido));
 	}
 }
